Relocate `From<LazyExpandedValue> for LazyValue` impls
