+++ src/lazy/expanded/mod.rs
@@ -47,9 +47,12 @@ use crate::lazy::encoding::RawValueLiteral;
 use crate::lazy::expanded::macro_evaluator::EExpEvaluator;
 use crate::lazy::expanded::macro_table::MacroTable;
 use crate::lazy::expanded::r#struct::LazyExpandedStruct;
+use crate::lazy::r#struct::LazyStruct;
 use crate::lazy::raw_stream_item::RawStreamItem;
 use crate::lazy::raw_value_ref::RawValueRef;
+use crate::lazy::sequence::{LazyList, LazySExp};
 use crate::lazy::str_ref::StrRef;
+use crate::lazy::value::LazyValue;
 use crate::raw_symbol_token_ref::AsRawSymbolTokenRef;
 use crate::result::IonFailure;
 use crate::{
@@ -303,6 +306,38 @@ impl<'top, 'data: 'top, D: LazyDecoder<'data>> LazyExpandedValue<'top, 'data, D>
 
     pub fn context(&self) -> EncodingContext<'top> {
         self.context
+    }
+}
+
+impl<'top, 'data, D: LazyDecoder<'data>> From<LazyExpandedValue<'top, 'data, D>>
+    for LazyValue<'top, 'data, D>
+{
+    fn from(expanded_value: LazyExpandedValue<'top, 'data, D>) -> Self {
+        LazyValue { expanded_value }
+    }
+}
+
+impl<'top, 'data, D: LazyDecoder<'data>> From<LazyExpandedStruct<'top, 'data, D>>
+    for LazyStruct<'top, 'data, D>
+{
+    fn from(expanded_struct: LazyExpandedStruct<'top, 'data, D>) -> Self {
+        LazyStruct { expanded_struct }
+    }
+}
+
+impl<'top, 'data, D: LazyDecoder<'data>> From<LazyExpandedSExp<'top, 'data, D>>
+    for LazySExp<'top, 'data, D>
+{
+    fn from(expanded_sexp: LazyExpandedSExp<'top, 'data, D>) -> Self {
+        LazySExp { expanded_sexp }
+    }
+}
+
+impl<'top, 'data, D: LazyDecoder<'data>> From<LazyExpandedList<'top, 'data, D>>
+    for LazyList<'top, 'data, D>
+{
+    fn from(expanded_list: LazyExpandedList<'top, 'data, D>) -> Self {
+        LazyList { expanded_list }
     }
 }
